requirements.txt
torch==1.8.1
torchvision==0.9.1
matplotlib==3.4.1
torchcam==0.2.0
spacy==3.0.6
sklearn==0.24.1
Pillow==8.3.1